category service/controller

+++ quiz-web.Server/Data/Entity/AnswerCount.cs
@@ -2,12 +2,12 @@
 {
     public class AnswerCount
     {
-        public long Id { get; set; }
-        public long QuestionId { get; set; }
-        public long CntOpt1 { get; set; }
-        public long CntOpt2 { get; set; }
-        public long CntOpt3 { get; set; }
-        public long CntOpt4 { get; set; }
+        public int Id { get; set; }
+        public int QuestionId { get; set; }
+        public int CntOpt1 { get; set; }
+        public int CntOpt2 { get; set; }
+        public int CntOpt3 { get; set; }
+        public int CntOpt4 { get; set; }
 
         public Question Question { get; set; }
     }
