+++ quiz-web.Server/Data/Entity/Purchase.cs
@@ -2,8 +2,8 @@
 {
     public class Purchase
     {
-        public long Id { get; set; }
-        public long QuizzId { get; set; }
+        public int Id { get; set; }
+        public int QuizzId { get; set; }
         public string UserId { get; set; }
         public long Price { get; set; }
         public DateTime PurchaseAt { get; set; }
