+++ quiz-web.Server/Data/Entity/Question.cs
@@ -2,7 +2,7 @@
 {
     public class Question
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Img { get; set; }
         public string Content { get; set; }
         public string Opt1 { get; set; }
@@ -10,8 +10,8 @@
         public string Opt3 { get; set; }
         public string Opt4 { get; set; }
         public string Correct { get; set; }
-        public long QuizzId { get; set; }
-        public long Ordinal { get; set; }
+        public int QuizzId { get; set; }
+        public int Ordinal { get; set; }
 
         public Quizz Quizz { get; set; }
         public ICollection<Result> Results { get; set; }
