+++ quiz-web.Server/Migrations/QDbContextModelSnapshot.cs
@@ -157,26 +157,26 @@ namespace quiz_web.Server.Migrations
 
             modelBuilder.Entity("quiz_web.Server.Data.Entity.AnswerCount", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<long>("CntOpt1")
-                        .HasColumnType("bigint");
+                    b.Property<int>("CntOpt1")
+                        .HasColumnType("int");
 
-                    b.Property<long>("CntOpt2")
-                        .HasColumnType("bigint");
+                    b.Property<int>("CntOpt2")
+                        .HasColumnType("int");
 
-                    b.Property<long>("CntOpt3")
-                        .HasColumnType("bigint");
+                    b.Property<int>("CntOpt3")
+                        .HasColumnType("int");
 
-                    b.Property<long>("CntOpt4")
-                        .HasColumnType("bigint");
+                    b.Property<int>("CntOpt4")
+                        .HasColumnType("int");
 
-                    b.Property<long>("QuestionId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("QuestionId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -226,8 +226,8 @@ namespace quiz_web.Server.Migrations
                     b.Property<string>("PasswordHash")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<long>("PaymentId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("PaymentId")
+                        .HasColumnType("int");
 
                     b.Property<string>("PhoneNumber")
                         .HasColumnType("nvarchar(max)");
@@ -266,17 +266,17 @@ namespace quiz_web.Server.Migrations
 
             modelBuilder.Entity("quiz_web.Server.Data.Entity.Attempt", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<DateTime>("AttemptAt")
                         .HasColumnType("datetime2");
 
-                    b.Property<long>("QuizzId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("QuizzId")
+                        .HasColumnType("int");
 
                     b.Property<string>("Score")
                         .IsRequired()
@@ -300,11 +300,11 @@ namespace quiz_web.Server.Migrations
 
             modelBuilder.Entity("quiz_web.Server.Data.Entity.Category", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<string>("Name")
                         .IsRequired()
@@ -317,11 +317,11 @@ namespace quiz_web.Server.Migrations
 
             modelBuilder.Entity("quiz_web.Server.Data.Entity.PaymentCard", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.HasKey("Id");
 
@@ -330,11 +330,11 @@ namespace quiz_web.Server.Migrations
 
             modelBuilder.Entity("quiz_web.Server.Data.Entity.Purchase", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<long>("Price")
                         .HasColumnType("bigint");
@@ -342,8 +342,8 @@ namespace quiz_web.Server.Migrations
                     b.Property<DateTime>("PurchaseAt")
                         .HasColumnType("datetime2");
 
-                    b.Property<long>("QuizzId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("QuizzId")
+                        .HasColumnType("int");
 
                     b.Property<string>("UserId")
                         .IsRequired()
@@ -360,11 +360,11 @@ namespace quiz_web.Server.Migrations
 
             modelBuilder.Entity("quiz_web.Server.Data.Entity.Question", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<string>("Content")
                         .IsRequired()
@@ -394,11 +394,11 @@ namespace quiz_web.Server.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<long>("Ordinal")
-                        .HasColumnType("bigint");
+                    b.Property<int>("Ordinal")
+                        .HasColumnType("int");
 
-                    b.Property<long>("QuizzId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("QuizzId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -409,21 +409,21 @@ namespace quiz_web.Server.Migrations
 
             modelBuilder.Entity("quiz_web.Server.Data.Entity.Quizz", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<long>("AttemptCounts")
-                        .HasColumnType("bigint");
+                    b.Property<int>("AttemptCounts")
+                        .HasColumnType("int");
 
                     b.Property<string>("AuthorId")
                         .IsRequired()
                         .HasColumnType("nvarchar(450)");
 
-                    b.Property<long>("CategoryId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("CategoryId")
+                        .HasColumnType("int");
 
                     b.Property<DateTime>("CreatedAt")
                         .HasColumnType("datetime2");
@@ -438,8 +438,8 @@ namespace quiz_web.Server.Migrations
                     b.Property<long>("Price")
                         .HasColumnType("bigint");
 
-                    b.Property<long>("QuestionCounts")
-                        .HasColumnType("bigint");
+                    b.Property<int>("QuestionCounts")
+                        .HasColumnType("int");
 
                     b.Property<string>("Title")
                         .IsRequired()
@@ -459,17 +459,17 @@ namespace quiz_web.Server.Migrations
 
             modelBuilder.Entity("quiz_web.Server.Data.Entity.Result", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<long>("AttemptId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("AttemptId")
+                        .HasColumnType("int");
 
-                    b.Property<long>("QuestionId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("QuestionId")
+                        .HasColumnType("int");
 
                     b.Property<string>("SelectedOption")
                         .IsRequired()
